add no cache to force theta binary download
add --no-cache to docker build such that if we rebuild the docker on the same machine as prior, it will force the curl download of the theta binaries. The theta binaries do not carry the version in the filename and it is assumed "latest" when we curl it. Updating to GN 3.4.0, yields the same curl download request, so we should force it here.